Rearrange statistics modules into classes. Create statistics class and put helper method into it.
