+++ lib/stat_tracker.rb
@@ -1,16 +1,11 @@
-require_relative './modules/game_statistics'
-require_relative './modules/league_statistics'
-require_relative './modules/season_statistics'
-require_relative './modules/team_statistics'
+require_relative 'game_statistics'
+require_relative 'league_statistics'
+require_relative 'season_statistics'
 require_relative 'game_team'
 require_relative 'game'
 require_relative 'team'
 
 class StatTracker
-  include GameStatistics
-  include LeagueStatistics
-  include SeasonStatistics
-  include TeamStatistics
 
   def self.from_csv(locations)
     game_path = locations[:games]
@@ -21,8 +16,29 @@ class StatTracker
 
   attr_reader :games, :teams, :game_teams
   def initialize(game_path, team_path, game_teams_path)
-    @games ||= Game.create_games(game_path)
-    @teams ||= Team.create_teams(team_path)
-    @game_teams ||= GameTeam.create_game_teams(game_teams_path)
+    @games = Game.create_games(game_path)
+    @teams = Team.create_teams(team_path)
+    @game_teams = GameTeam.create_game_teams(game_teams_path)
+    @game_statistic = GameStatistics.new
+    @league_statistic = LeagueStatistics.new
+    @season_statistic = SeasonStatistics.new
   end
+
+  def highest_total_score; @game_statistic.highest_total_score; end
+
+  def lowest_total_score; @game_statistic.lowest_total_score; end
+
+  def biggest_blowout; @game_statistic.biggest_blowout; end
+
+  def count_of_games_by_season; @game_statistic.count_of_games_by_season; end
+
+  def average_goals_per_game; @game_statistic.average_goals_per_game; end
+
+  def average_goals_by_season; @game_statistic.average_goals_by_season; end
+
+  def percentage_home_wins; @game_statistic.percentage_home_wins; end
+
+  def percentage_visitor_wins; @game_statistic.percentage_visitor_wins; end
+
+  def percentage_ties; @game_statistic.percentage_ties; end
 end
