+++ test/stat_tracker_test.rb
@@ -5,6 +5,10 @@ require './lib/stat_tracker'
 require './lib/game'
 require './lib/team'
 require './lib/game_team'
+require './lib/game_statistics'
+require './lib/league_statistics'
+require './lib/season_statistics'
+
 
 class StatTrackerTest < Minitest::Test
   def setup
